move a bunch of stuff into geocoding helper
